requirements.txt
mcp>=1.0.0
mcp-neo4j-cypher>=0.5.1
neo4j>=5.15.0
python-dotenv>=1.0.0
rich>=13.7.0
requests>=2.31.0
